refactor: 로또 생성 개수 자료형 Int -> Long
- 금액의 자료형과 통일성을 위해 Long으로 변경

+++ src/main/java/controller/LottoController.java
@@ -20,7 +20,7 @@ public class LottoController {
         long money = inputView.inputMoney();
         lottoService = new LottoService(money);
 
-        int times = lottoService.getTimes();
+        long times = lottoService.getTimes();
         outputView.putTimes(times);
 
         if (times < 1) {
