+++ src/test/java/service/LottoServiceTest.java
@@ -16,7 +16,7 @@ public class LottoServiceTest {
     @DisplayName("금액에 대해 알맞은 로또 구매 갯수를 추출한다.")
     void lottoGroupSizeTest(long money, int expectedTimes) {
         LottoService lottoService = new LottoService(money);
-        int times = lottoService.getTimes();
+        long times = lottoService.getTimes();
         assertThat(times).isEqualTo(expectedTimes);
     }
